Move slashValidator to BeaconStateMutators
Reorganize logic to support this change.
