+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/logic/versions/altair/util/BlockProcessorAltair.java
@@ -36,6 +36,7 @@ import tech.pegasys.teku.spec.datastructures.state.beaconstate.MutableBeaconStat
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.versions.altair.MutableBeaconStateAltair;
 import tech.pegasys.teku.spec.datastructures.type.SszPublicKey;
 import tech.pegasys.teku.spec.logic.common.block.AbstractBlockProcessor;
+import tech.pegasys.teku.spec.logic.common.helpers.BeaconStateMutators;
 import tech.pegasys.teku.spec.logic.common.helpers.Predicates;
 import tech.pegasys.teku.spec.logic.common.operations.validation.AttestationDataStateTransitionValidator;
 import tech.pegasys.teku.spec.logic.common.statetransition.exceptions.BlockProcessingException;
@@ -61,6 +62,7 @@ public class BlockProcessorAltair extends AbstractBlockProcessor {
       final Predicates predicates,
       final MiscHelpersAltair miscHelpers,
       final BeaconStateAccessorsAltair beaconStateAccessors,
+      final BeaconStateMutators beaconStateMutators,
       final BeaconStateUtil beaconStateUtil,
       final AttestationUtil attestationUtil,
       final ValidatorsUtil validatorsUtil,
@@ -70,6 +72,7 @@ public class BlockProcessorAltair extends AbstractBlockProcessor {
         predicates,
         miscHelpers,
         beaconStateAccessors,
+        beaconStateMutators,
         beaconStateUtil,
         attestationUtil,
         validatorsUtil,
@@ -152,11 +155,11 @@ public class BlockProcessorAltair extends AbstractBlockProcessor {
       if (shouldUpdate) {
         epochParticipation.set(index, SszByte.of(participationFlags));
 
-        final int proposerIndex = beaconStateUtil.getBeaconProposerIndex(state);
+        final int proposerIndex = beaconStateAccessors.getBeaconProposerIndex(state);
         final UInt64 proposerReward =
             proposerRewardNumerator.dividedBy(
                 specConfig.getProposerRewardQuotient().times(WEIGHT_DENOMINATOR));
-        validatorsUtil.increaseBalance(state, proposerIndex, proposerReward);
+        beaconStateMutators.increaseBalance(state, proposerIndex, proposerReward);
       }
     }
   }
@@ -229,16 +232,17 @@ public class BlockProcessorAltair extends AbstractBlockProcessor {
             .max(specConfig.getEffectiveBalanceIncrement());
 
     UInt64 proposerReward = UInt64.ZERO;
-    final int beaconProposerIndex = beaconStateUtil.getBeaconProposerIndex(state);
+    final int beaconProposerIndex = beaconStateAccessors.getBeaconProposerIndex(state);
     for (Integer includedIndex : includedIndices) {
       final UInt64 effectiveBalance = validators.get(includedIndex).getEffective_balance();
       final UInt64 inclusionReward =
           maxSlotRewards.times(effectiveBalance).dividedBy(committeeEffectiveBalance);
       UInt64 proposerShare = inclusionReward.dividedBy(specConfig.getProposerRewardQuotient());
       proposerReward = proposerReward.plus(proposerShare);
-      validatorsUtil.increaseBalance(state, includedIndex, inclusionReward.minus(proposerShare));
+      beaconStateMutators.increaseBalance(
+          state, includedIndex, inclusionReward.minus(proposerShare));
     }
-    validatorsUtil.increaseBalance(state, beaconProposerIndex, proposerReward);
+    beaconStateMutators.increaseBalance(state, beaconProposerIndex, proposerReward);
   }
 
   private boolean eth2FastAggregateVerify(
